+++ app/charts/shared/chart-data-filters.tsx
@@ -16,7 +16,7 @@ import {
   OptionGroup,
   Option,
 } from "@/configurator";
-import { hierarchyToOptions, TimeInput } from "@/configurator/components/field";
+import { TimeInput } from "@/configurator/components/field";
 import {
   getTimeIntervalFormattedSelectOptions,
   getTimeIntervalWithProps,
@@ -35,6 +35,7 @@ import {
 } from "@/graphql/query-hooks";
 import { Icon } from "@/icons";
 import { useLocale } from "@/locales/use-locale";
+import { hierarchyToOptions } from "@/utils/hierarchy";
 
 export const ChartDataFilters = ({
   dataSet,